WIP - updates to the demodulator viewer app and the recently used files menu.
